refactor: change makeRemote prefix to only make on factories about usecase to avoid overcomplexity name
