+++ pages/index.tsx
@@ -2,7 +2,7 @@ import { GetServerSideProps } from "next"
 
 import { BookList, Header } from "@/presentation/components"
 import { BookModel } from "@/domain/models"
-import { makeRemoteLoadBooks } from "@/main/factory/usecases"
+import { makeLoadBooks } from "@/main/factory/usecases"
 
 type HomePageProps = {
     books: BookModel[]
@@ -19,7 +19,7 @@ function HomePage({ books }: HomePageProps) {
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
     const accessToken = context.req.cookies['bookue-user']
-    const remoteLoadBooks = makeRemoteLoadBooks()
+    const remoteLoadBooks = makeLoadBooks()
     const books = await remoteLoadBooks.loadBooks({ accessToken })
     return {
         props: {
