+++ src/main/factory/usecases/add.account.factory.ts
@@ -2,4 +2,4 @@ import { RemoteAddAccount } from '@/data/usecases/remote.add.account';
 import { AddAccount } from '@/domain/usecases';
 import { makeAxiosHttpClient, makeBookueApiUrl } from '../http';
 
-export const makeRemoteAddAccount = (): AddAccount => new RemoteAddAccount(makeBookueApiUrl(''), makeAxiosHttpClient());
+export const makeAddAccount = (): AddAccount => new RemoteAddAccount(makeBookueApiUrl(''), makeAxiosHttpClient());
