+++ src/main/factory/usecases/add.note.factory.ts
@@ -1,13 +1,13 @@
-import { RemoteDeleteNote } from '@/data/usecases';
-import { DeleteNote } from '@/domain/usecases';
+import { RemoteAddNote } from '@/data/usecases';
+import { AddNote } from '@/domain/usecases';
 import { makeAxiosHttpClient, makeBookueApiUrl } from '../http';
 import { CookieManagerAdapter } from '@/infra/cookie';
 
-export const makeRemoteDeleteNote = (): DeleteNote => {
+export const makeAddNote = (): AddNote => {
     const apiUrl = makeBookueApiUrl('')
     const axiosHttpClient = makeAxiosHttpClient()
     const cookieManagerAdapter = new CookieManagerAdapter()
-    return new RemoteDeleteNote(
+    return new RemoteAddNote(
         cookieManagerAdapter,
         apiUrl,
         axiosHttpClient,
