+++ src/main/factory/usecases/authentication.factory.ts
@@ -2,4 +2,4 @@ import { RemoteAuthentication } from '@/data/usecases';
 import { Authentication } from '@/domain/usecases';
 import { makeAxiosHttpClient, makeBookueApiUrl } from '../http';
 
-export const makeRemoteAuthentication = (): Authentication => new RemoteAuthentication(makeBookueApiUrl(''), makeAxiosHttpClient());
+export const makeAuthentication = (): Authentication => new RemoteAuthentication(makeBookueApiUrl(''), makeAxiosHttpClient());
