+++ src/main/factory/usecases/delete.note.factory.ts
@@ -1,15 +1,15 @@
-import { RemoteAddBook } from '@/data/usecases';
-import { AddBook } from '@/domain/usecases';
+import { RemoteDeleteNote } from '@/data/usecases';
+import { DeleteNote } from '@/domain/usecases';
 import { makeAxiosHttpClient, makeBookueApiUrl } from '../http';
 import { CookieManagerAdapter } from '@/infra/cookie';
 
-export const makeRemoteAddBook = (): AddBook => {
+export const makeDeleteNote = (): DeleteNote => {
     const apiUrl = makeBookueApiUrl('')
     const axiosHttpClient = makeAxiosHttpClient()
     const cookieManagerAdapter = new CookieManagerAdapter()
-    return new RemoteAddBook(
-        apiUrl,
+    return new RemoteDeleteNote(
         cookieManagerAdapter,
+        apiUrl,
         axiosHttpClient,
     )
 }
