+++ src/main/factory/usecases/index.ts
@@ -1,8 +1,8 @@
-export * from './remote.add.account.factory';
-export * from './remote.add.book.factory';
-export * from './remote.add.note.factory';
-export * from './remote.authentication.factory';
-export * from './remote.load.books.factory';
-export * from './remote.load.book.factory';
-export * from './remote.load.notes.factory';
-export * from './remote.delete.note.factory';
+export * from './add.account.factory';
+export * from './add.book.factory';
+export * from './add.note.factory';
+export * from './authentication.factory';
+export * from './load.books.factory';
+export * from './load.book.factory';
+export * from './load.notes.factory';
+export * from './delete.note.factory';
