+++ src/main/factory/usecases/load.book.factory.ts
@@ -2,4 +2,4 @@ import { LoadBook } from '@/domain/usecases';
 import { RemoteLoadBook } from '@/data/usecases';
 import { makeAxiosHttpClient, makeBookueApiUrl } from '../http';
 
-export const makeRemoteLoadBook = (): LoadBook => new RemoteLoadBook(makeBookueApiUrl(''), makeAxiosHttpClient());
+export const makeLoadBook = (): LoadBook => new RemoteLoadBook(makeBookueApiUrl(''), makeAxiosHttpClient());
