+++ src/main/factory/usecases/load.books.factory.ts
@@ -2,4 +2,4 @@ import { LoadBooks } from '@/domain/usecases';
 import { RemoteLoadBooks } from '@/data/usecases';
 import { makeAxiosHttpClient, makeBookueApiUrl } from '../http';
 
-export const makeRemoteLoadBooks = (): LoadBooks => new RemoteLoadBooks(makeBookueApiUrl(''), makeAxiosHttpClient());
+export const makeLoadBooks = (): LoadBooks => new RemoteLoadBooks(makeBookueApiUrl(''), makeAxiosHttpClient());
